+++ puffin_egui/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the egui crate will be documented in this file.
 ## Unreleased
 * Show total frames recorded and their total size.
 * In-memory compression of frames to use up less RAM.
+* Add slider to control how many recent frames to store.
 
 
 ## 0.10.3 - 2021-11-08